move event ipc handling logic to events controller class
